requirements.txt
discord.py
requests
colorama
cachetools
